Anchor and Target Movement Mostly Complete
The movement for Anchors to a central point over a certain amount of
time is done.  The point can move and the anchors will follow it.
Anchors do not yet despawn when they hit the point.

Targets rotate around anchors roughly on a plane perpendicular to the
vector towards the player.  It's too late and I don't have enough
thinking power to figure out the method for correcting the trajectory so
the plane it is rotating on also intersects the anchor.