Fixed Animation Syncing and Edited Gun Script
Edits to Gun and GunAnimations script, did my testing and most edits in
throwaway script called TempGunForTesting. It's a version without all
the NVR, but only difference when with NVR will be inputs.  I commented
out Magazine and MagCollider so I would stop getting errors with
dependencies it had with Gun script. The gun changes should've been in
seperate branch, but was already far in when noticed. I'll branch off
and pull this into there.